Smaller improvements in re release procedure. Phpstan latest version fixed.
git-svn-id: https://svn.viathinksoft.com/svn/oidplus/trunk@1696 02e4d621-2042-4998-a0bc-9ccd24201011

+++ plugins/viathinksoft/publicPages/095_attachments/OIDplusPagePublicAttachments.class.php
@@ -441,7 +441,7 @@ class OIDplusPagePublicAttachments extends OIDplusPagePluginPublic
 	 */
 	private static function convert_filesize(int $bytes, int $decimals = 2): string {
 		$size = array(_L('Bytes'),_L('KiB'),_L('MiB'),_L('GiB'),_L('TiB'),_L('PiB'),_L('EiB'),_L('ZiB'),_L('YiB'));
-		$factor = floor((strlen("$bytes") - 1) / 3);
+		$factor = (int)floor((strlen("$bytes") - 1) / 3);
 		return sprintf("%.{$decimals}f", $bytes / pow(1024, $factor)) . ' ' . @$size[$factor];
 	}
